+++ src-ui/src/app/components/admin/logs/logs.component.html
